Enhance: more information available from TLuaInterpreter::startLogging()
Reformat error message to current proposed "standard" and make it capable
of being internationalised.

Now return a Boolean on success or failure to start or stop logging and an
(internationaliseable) message indicating what happened, if anything and
including the log file-name if applicable.

In light of the messages now available to the lua command the code has been
revised so it will no longer cause a logging starting/stopping system
message to appear on the main console - though using the button on the main
console will continue to do so.)

The existing log-to-file button used an icon that appears to have been an
old Oxygen "folder-downloads.png" icon image and the only way to tell if it
was active was a system dependent colouration that "on" widgets adopt.  I
have taken the current 48x48 Oxygen "text-x-log.png" and superimposed
smaller (32x32) "media-record.png" and "media-stop.png" to create two
similar in style but visibly different "log-record" and "log-stop" icons,
they, like any other icons I have produced from the Oxygen icon set, are
released under a GPLv2+ licence for use in Mudlet and anywhere else that
is permitted.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>