BugFix: Prevent colouring of room exit number tool-tips on room exit dialog
As referred to in https://bugs.launchpad.net/mudlet/+bug/1369040 it has
been found that colouration applied to the room numbers displayed in the
map room exits dialog to indicate their validity was being passed onto
the tool-tips for those elements. The cure is to explicitly apply the
colouration to ONLY the QLineEdits involved.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>