Enhance: more information available from TLuaInterpreter::startLogging()
Reformat error message to current proposed "standard" and make it capable
of being internationalised.

Now returns: a Boolean on success or failure to start or stop logging; an
(internationaliseable) message indicating what happened, if anything and
including the log file-name if applicable; the log file-name on its own
and finally a numeric code indicating the result (for script writers
working for an international audience once Mudlet goes multi-lingual!)

In light of the messages now available to the lua command the code has been
revised so it will no longer cause a logging starting/stopping system
message to appear on the main console - though using the button on the main
console will continue to do so.)

The existing log-to-file button used an icon that appears to have been an
old Oxygen "folder-downloads.png" icon image and the only way to tell if it
was active was a system dependent colouration that "on" widgets adopt.  I
have taken that icon and placed a red cross over it to indicate the state
of logging being active and that pressing the button will stop logging.
This "folder-downloads-red-cross.png", like any other icons I have produced
from the Oxygen icon set or otherwise, it is released under a GPLv2+
licence for use in Mudlet and anywhere else permitted.

Also added my details to copyrights of touched files from this and a
previous Pull-Req.

Remove a pair of duplicated setFocusPolicy() calls on the
TConsole::mpBufferSearchUp & Down members, and a
TConsole::setAttribute( Qt::WA_OpaquePaintEvent ) call.

Corrected the tool-tips for the search up and down buttons - both had
"next result".  Also included "<html><head/><body><p>" and
"</p></body></html>" wrappers around all the tool tips because this causes
them to be wrapped to fit the display in an optimum manner.  Also pass them
through a call to tr() so they can be marked for translation.  I WOULD
URGE OTHER CODERS TO ENSURE THAT TOOL-TIPS IN AREAS THEY WORK ON - BOTH
IN .CPP CODE AND IN .UI FORMS - ARE REVISED TO USE THIS FORMAT.

*** THE CODE PRESENT IN THIS BRANCH OF MY PUBLIC REPOSITORY HAS BEEN
RE-BASED TO FOLLOW CHANGES IN THE MAIN CENTRAL REPOSITORY "development"
BRANCH AFTER THE PREVIOUS COMMITS WERE PUBLISHED - THIS MIGHT MESS THINGS
UP - BUT I HOPE NOT! 8-(  ***

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>