Add the new "sysNegotiationFinishedEvent"
The event is fired when one of the data protocols MSDP, MXP, ATCP, GMCP or
Channel 102 were successfully enabled. User scripts can now freely use the
protocol.

The event has a second argument that denominates the protocol that was enabled.

The commit also makes sure that GMCP is only enabled if the user chose to enable
GMCP if the server sends IAC DO GMCP.