Revise: modified lua function(s) ret. values methodology to agreed standard
Due to unawareness of standards I have deviated from them in producing
multiple return values upon success as well as failure.  Whilst I feel that
a second return value being a translatable status string and a third being
an integer encoding of the same might have some merit, particularly in a
fully internationalised and localised version of Mudlet in the future; no
decision has been made over precisely how I18n and L10n will be handled so
using them is a little premature and only confuses things for the moment.

The current methodology is:
* For bad argument TYPE situations: to throw a lua_error and provide a
error message using a fairly uniform construction, I provide these via
Qt's translation system to permit that part of the Lua API to be
translated with the C++ parts of the application's framework.

* For bad argument VALUE situations best practice is for a lua function to
provide a primary nil return value but to also provide a secondary value as
a string message describing the (first) argument value error.  A lua error
is NOT thrown but unless the user catches the error message by providing a
second variable to contain the second return value they may not find out
what is wrong.  Again I choose to provide a translatable message for this.

* For good arguments case that are satisfactory but produce no data, an
empty table, an empty string or a signal number value is an appropriate
primary return value. For example for either type of table that the lua
getAreaExits((int) areaId, <(bool) provide Full data> ) function now
returns, an empty table will be returned for an area which is isolated and
has no exits.  A secondary return value is NOT expected in this situation
as far as I can tell, though I believe one could be useful...

* For good arguments values for functions that provide data then the data
is returned as the primary and only return value.

* For good arguments values for functions that do not return data then
success should be indicated by a single true Boolean return value.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>