BugFix: Treat and prevent further duplicate or empty map area names
Some steps had previously been taken to enforce no duplicate area names
when importing an XML map or when adding a new name but other mechanisms
were left unchecked, specifically this was when renaming an existing area.

This commit adds a number of validation steps to many of the lua functions
that manipulate areas and their names.  This includes taking steps when
creating a new TArea instance to ensure a suitable area name is created and
added to the areaNamesMap which may be overwritten if a valid (non-empty,
non-duplicate) name is provided at the time or later.  This avoids problems
with the Area selection widget on the 2 or 3D Map Display and the lua
getAreaTable() function, neither of which will handle duplicate area names
and for the former, does not handle nameless areas well either.  The
deleteArea() function could take either an area Id or a name as the target
of its action so the area that is to be deleted is now not subject to any
ambiguity if supplied as a name!

Should a map file be loaded where empty or duplicated area names are found
these will be "fixed" and a warning dialog {with the mudlet_warning.png
icon I created a while back but has not yet been included from other
outstanding Pull Request 8-) } raised to explain what has happened.  The
user will only get this once per map file as there should now be no way to
modified the map file to have either of these issues.

The code that builds the area selection widget is revised to handle the
corner cases of two areas that have the same letters in their name but the
cases vary (the widget is sorted by name in a case insensitive manner)
which previously was not handled (same as duplicate names were not).

As a side effect of revising the TLuaInterpreter Class, area names
containing non-ASCII characters can now be handled - they will be passed
through the lua subsystem using the UTF-8 encoding.

*** This commit has been re-based and the branch it was in on my Repository
was effectively renamed - there originally were two versions of the commit
which applied to the development branch before and after a different one
was to be applied.  The branch names reflected this but since the different
branch "fix_AreaExits" has now been applied this version which originally
was entitled "fix_areaExits_&_duplicateAreaIssue" was retitled and
overwrote the original "fix_duplicateAreaIssue"! ***

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>