fixup: remove cruft a.k.a. dead stuff that add nothing to code
Takes out "dlgOptionsAreaXXX" header & class file and
"options_area_XXXX.ui""ui forms (the latter contain a single QWidget) for
each of Actions (a.k.a. "Buttons"), Aliases, Scripts, Timers and Triggers,
none of which are actually used for anything.

In mudlet::mudlet():
* Take out lines that enable some toolbar buttons on main toolbar at startup
as the buttons are immediately disabled by call to
mudlet::disableToolbarButtons() afterwards.

* Take out the second of each of two doubled up lines of code where
repetition does not alter things.

In src.pro project files for non-MSVC builds:
* Take out any optimisation {-Ox} flags that Qt puts in automagically for
RELEASE builds as they are overridden by our choices but still show up
during compilation which is a little confusing.

* Take out the -g flag from the DEBUG build flags as Qt automatically
inserts it for us for those type builds, and the duplicate shown during
compilation because we also add it as well, is not helpful.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>