FixEnhance: fix crashes on room deletion, speed up room deletion
Prior code to speed up room deletion used a global QMultiHash to store,
using room Ids as keys and values the rooms (values) that have one or more
exits that lead into the given room (key).  Whilst this was adjusted for
new rooms being added or any rooms being deleted, it was not generated when
a map was loaded and not amended if a room's exits were changed by Lua
command or by Map GUI.  If this data was correctly maintained, when a room
was deleted it was still necessary to scan through all the exits to find
the particular one(s) that lead to the room being deleted.

The replacement code here instead stores the "Entrance" data in the room
to which it applies.  For compatibility with existing maps (and efficiency)
the data is generated at the same time the exits are validated in the
TRoom::auditExits() code though provision is made so that when the map
file format is increment (to 17) the data will be saved with and restored
from the map file.  To maintain the data the action of setting an exit
for a room will also set the corresponding "entrance" in the room on the
other end of the route.  The entrance data is held within two PRIVATE
QSets:
TRoom::QSet<QPair<quint8, quint64> > mNormalEntrances
TRoom::QSet<QPair<QString, quint64> > mSpecialEntrances;
which avoids issues of duplication that the QList type present.  The room
Ids of the rooms that have exits to THIS room are stored as quint64 - as
negative room Ids are not permitted and to use the largest integral value
type Qt supports.  The direction component represents the direction to use
FROM the other room to get to THIS room and is an unsigned byte for Normal
exits and the QString value for Special ones, the latter WITHOUT the '0'
or '1' prefix that is (currently!?) used to record exit lock details.

The entrance data may be of use to mapper scripts and users, so provision
to gain read access to it is provided via two new getter TRoom methods
(two setter and two resetter methods also added but for safety they are
kept private, two further ones needed by other map related classes are
exposed but they take the QPair data directly).  However this data awaits
the feature to be release in the Lua subsystem!

In doing this I became aware that:
* there was not comparable getters for the Exits - but it was simple to
  put them together.
* there was a multitude of individual direction exit setters - to keep the
  setting code straightforward it was simpler to replace them with a single
  one that took a second, directional argument.

It is intended to give access to the user to this data but as that is not
a requirement / desirable for the target of this commit it will follow as
a separate "feature" to add later.

Adds:
(public):
* (QSet<QPair<quint8,quint64>>)TRoom::getNormalExits()
* (QSet<QPair<QString,quint64>>)TRoom::getSpecialExits()
* (QSet<QPair<quint8,quint64>>)TRoom::getNormalEntrances()
* (QSet<QPair<QString,quint64>>)TRoom::getSpecialEntrances()

(private):
* (QSet<QPair<quint8,quint64>>)TRoom::mNormalEntrances
* (QSet<QPair<QString,quint64>>)TRoom::mSpecialEntrances

* TRoom::setEntrance( fromRoomId, fromDirectionCode ) {Normal routes}
* TRoom::setEntrance( fromRoomId, fromDirectionString ) {Special routes}
* TRoom::resetEntrance( fromRoomId, fromDirectionCode ) {Normal routes}
* TRoom::resetEntrance( fromRoomId, fromDirectionString ) {Special routes}

Modifies:
* (void)TRoom::restore((QDataStream &),(int)room,(int)version)
  to (void)TRoom::restore((QDataStream &),(int)version)
* (void)TRoomDB::restoreSingleRoom((QDataStream &),(int)roomID,(TRoom *))
  to (void)TRoomDB::restoreSingleRoom((QDataStream &),(TRoom *))

Removes:
* (int)TMap::mPlausaOptOut - just not used
* (int)TRoom::v - ditto
* (QMultiHash<int,int>)TRoomDB::reverseExitMap - recent addition but code
    to use it incomplete.  Also the use of the words "reverse exit" is
    misleading - there is no certainty in most MUDs that a route from one
    room to another has an identical opposite, this is _usually_ true but
    need not be so.  Using the word "Entrance" more clearly denotes the
    opposite state to "Exit"...!
* TRoom::getExits() - recent addition but code to use it incomplete.
    New code uses getNormalExits() and getSpecialExits() with more
    useful return types.
* TRoom::removeAllSpecialExitsToRoom(int) - code revisions make it
    redundant.
* TRoom::set{North|NorthEast|...|In|Out}(toRoomID) - replaced by general
    setExit(toRoomId, directionCode) which handles setting up Entrance
    details in other room(s) {both the given room and any previous room in
    the given direction}.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>