Revise: add compatibility option to work around issues of map format change
For the previous changes to work best it was necessary to modify the Map
File Format to store some additional data about each room.  If this is
not done then the data has to be recalculated each time the map is loaded
which will introduce delays the user will notice.  This commit adds an
option on the last tab of the profile preferences to overcome the choice
that the code now makes about the Map file format to use based upon
whether the code is a "release" or a "development" build.

Also: it applies a HTML formatting wrapper to all the tool-tips actually on
the profile preferences dialog which improves the visual presentation -
they become dynamically and sensibly word-wrapped to fit the actual screen
upon which they appear - which I have recommended for tool-tips throughout
this project!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>