Fixup: add getAllRoomEntrances() to Lua command set
Though suitable for release code this was added to help
with debugging.  This was what enabled me to spot the
problems that the previous pair of commits ameliorates.

This currently only reports the rooms that have exit(s)
that lead to the given room Id.  It is anticipated that
there will be a future revision to report the particular
direction(s) from the given room(s) are the one(s) that
lead to the room, using a second argument that will be
a boolean true to trigger that behavior (the absence of,
or a false, second argument would then cause the result
that this code produces.)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>