FixEnhance: fix entranceMap maintenance, bulk room deletion and map loading
Previously we were not removing entries from the entranceMap involving
the value (a room that the room Id that was a key had an entrance FROM)
when a route was changed.  There is a performance cost in ensuring the
data is kept correctly - there may be a modest gain by storing the
entrance data within each TRoom class instance rather than a central
database in TRoomDB...

Deletion of multiple rooms and map loading can be done more efficiently
if we skip some redundant steps.

Also added/revised some timing code to measure things.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>