BugFix: Treat and prevent further duplicate or empty map area names
Some steps had previously been taken to enforce no duplicate area names
when importing an XML map or when adding a new name but other mechanisms
were left unchecked, specifically this was when renaming an existing area.

This commit adds a number of validation steps to many of the lua functions
that manipulate areas and their names.  This includes taking steps when
creating a new TArea instance to ensure a suitable area name is created and
added to the areaNamesMap which may be overwritten if a valid (non-empty,
non-duplicate) name is provided at the time or later.  This avoids problems
with the Area selection widget on the 2 or 3D Map Display and the lua
getAreaTable() function, neither of which will handle duplicate area names
and for the former, does not handle nameless areas well either.  The
deleteArea() function could take either an area Id or a name as the target
of its action so the area that is to be deleted is now not subject to any
ambiguity if supplied as a name!

Should a map file be loaded where empty or duplicated area names are found
these will be "fixed" and warning messages inserted onto the main profile
to explain what has happened.  The user will only get this once per map
file as there should now be no way to modified the map file to have either
of these issues.

The code that builds the area selection widget is revised to handle the
corner cases of two areas that have the same letters in their name but the
cases vary (the widget is sorted by name in a case insensitive manner)
which previously was not handled (same as duplicate names were not).

As a side effect of revising the TLuaInterpreter Class, area names
containing non-ASCII characters can now be handled - they will be passed
through the lua subsystem using the UTF-8 encoding.

*** This commit is a reworking of one that produced a QMessageBox to alert
and advised the user what was happening - that was deemed to be too
intrusive so this version instead displays the information in the console
- as such the detail of the renaming has to be shown now whereas the dialog
solution had the option of providing it as "Show Details..." to display it
only if requested at the time. ***

Additionally, the lua setAreaName(areaId, newAreaName) has been extended to
allow the existing area to be specified as a name (string) as well as an Id
as we can now uniquely identify it by that means.  This now matches the
behaviour of deleteArea which already acts in that manner.  It also means
a user script can use "setAreaNAme(oldName, newName)" to rename an area
without concern about determining the area Id.

This commit is based upon modifications made to the "development" branch
but corrects a bug that prevented setAreaName() from accepting a new
non-Ascii area name and includes the changes from a separate commit that
clears the area name table on loading a new map that was missed from the
original series of commits on the development branch.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>