bugFix: errors in Duplicate Area Issues Fix
When porting the code in Pull Requests #237 and #245 into a different
branch I spotted that TLuaInterpreter::setAreaName() would NOT handle
a new non-ASCII area name.  Also the code to update the area selection
widget in the mapper was being updated BEFORE the new area was added
in TLuaInterpreter::addAreaName() so it was not showing the new area until
another change was made that causes the widget to be updated.

In fixing the new area name handling I have imposed a new check that strips
off leading and trailing spaces from any argument given as a new area name
for both the above functions - this is to avoid breakages that might result
from an area name that is just a load of spaces - which would be confusing
in the area selection widget if permitted (it would be hard to
differentiate between "   " and "       " for instance!)

Also spotted a location in the misnamed TLuaInterpreter::getRoomAreaName()
where an area name was not being handled as a Utf8 encoded string, so
corrected that (and updated the error message formatting to current
standards).

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>