BugFix: Restore ability to create area name entry without a TArea instance
Before some recent changes it was possible to create a new area name entry
with TLuaInterpreter::setAreaName(...) by supplying an unused area ID and
a name so that it was stored in TRoomDB::areaNamesMap WITHOUT an actual
TArea instance to pair it with.  This fix restores that mode of operation
and the ability to DELETE such an orphan area name which had also been
removed by either the same series of commits or some other recent one that
neglected to continue to provide the latter.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>