BugFix: allow setRoomArea() to set room to area that has name but no TArea
After restoring the ability to register a (new) area ID with a specific
name in TRoomDB::areaNameMap without actually creating a TArea instance to
hold the area's data by using TLuaIntrepreter::setAreaName(areaId, name),
it is necessary to modify TLuaIntrepreter::setRoomArea( roomId, areaId) to
work again when "areaId" only represents an area Name without the TArea
to contain most of the area data.  This commit implements the changes to
do so - and permits an area Name to be used to specify the area instead of
an areaId provided such an area exists already.

Also arranges for the 2/3D mappers to be updated on successfully setting or
resetting a room's area - so the user can see the result on screen
immediately without having to cause the mapper to be redrawn.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>