BugFix: remove dlgTriggerEditor::accept() usage/properly escape a RegExp
Fixes a couple of one-liner errors that I recently introduced into the
Development branch (and also the release_30 but that uses different
commit references):

In commit-10fefa0a "revise: deletion of more redundant methods in header"
which removed definition of the above signal I forgot to remove the call
to connect(...) that used the signal.

In commit-94b492c1 "BugFix: Treat and prevent further duplicate or empty
map area names" I neglected to properly escape (using double '\') a regular
expression used to detect a de-duplicated Area Name.

Also: adds a couple of missing header files to qmake project file - their
absence means that they were NOT included in the group of files considered
when using the Qt IDE find/replace parts of the code editor - which can cause
problems when looking for usage of problematic functions/variables!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>