enhanceAndFix: fix handling of parallel edges in BGL route-finding graph
This commit cleans up some problems in route finding - specifically when
working out the exits to use to move from the ordered list of rooms that
make up the route that the A* code determined - no longer is the first
exit in a fixed order in a particular room used that happens to go to the
next room - instead the correct (lowest weighted, non-locked) exit is
selected.  If a route is not found the reason why is output via qDebug()
which should make debugging problems a bit easier for developers!  The BGL
graph that TMap::initgraph() producess now tracks which exit from one room
to another is the best one to use (choosing just one from equally weighted
ones and ignores higher weighted ones) if it is a "normal" exit the
subsequent use of TMap:findPath() will now insert a "translated" direction
to the TMap::mDirList - but in future it should be straightforward to
arrange for a user specified "direction" name to be used - perhaps with
locale specific defaults which will make using Mudlet's route-finding
code possible with non-ASCII using MUDs (i.e. those not based on American
English!)

This commit also avoids creating spurious entries in the BGL graph that
the previous code did by using the "[]" operator (which inserted *wrong*
entries into (QMap<int, int>) TMap::roomidToIndex for TRooms that should
NOT have been present!)  That was found to be causing crashes in corner
cases where ALL rooms in the map or at least some rooms or their exits
in what would otherwise have been a valid route were locked against use
for route-finding.

Also:
* Culls some dead code in TAstar.h that was copied verbatim from a usage
  example elsewhere (possible the boost library documentation).

* Removes following unused members of TMap:
  * (QList<int>) mTestedNodes
  * (QList<int>) conList
  * (int) mPlausaOptOut;
  * typedef of: mygraph_t::vertex_iterator to: vertex_iterator
  * typedef of: std::pair<int, int> to: edge

* Comments out the now unused members of TMap:
  * (QMap<int, int>) indexToRoomid

* Corrects spelling of TAstar.h in qmake project file - was spelt TAStar.h!

* Generates a "speedWalkWeight" table alongside the "speedWalkPath" and
  "speedWalkDir" ones that indicates the weight of each step in a route,
  it can be accessed in the same manner as those others.

* Host::assemblePath() now returns the total "weight" of the route
  calculated - which the lua command getPath(startRoomId, targetRoomId)
  now returns as a second value - this will make it easier for users'
  scripts to choose between routes to multiple destinations.

* Some Lua helper methods:
  * TLuaInterpreter::get_lua_string(...)
  * TLuaInterpreter::set_lua_string(...)
  * TLuaInterpreter::set_lua_table(...)
  now process strings as Utf-8 - this change was needed to allow non-ASCII
  characters to appear in exit directions but will also be needed by other
  commands. Whilst the first is currently unused, the second is used once
  in three class and the third by Host::assemblePath() and twice in
  TLuaInterpreter::initLuaGlobals() to set up something for "atcp" and
  "channel102" processing...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>