Enhance: Reformat Room Exit dialog tool-tips to use HTML formatting
It has been found that using HTML formatting for tool-tips improves their
on-screen formatting in that they are word wrapped to fit the available
on-screen space automatically rather than depending on manual layout by
the developer who created the message.  This commit applies this feature
to the room exits dialog.  The code has been done such a way as to minimise
the required HTML tags from the strings that a translator will have to
translate to support other languages.

ALSO:
* The tool-tip for the display of the room weight has been revised to
emphasise that it is used by default for exits that lead TO the room!

* In support of the previous detail the tool-tips that give the destination
room for the normal exits' room Ids QLineEdit boxes now also report the
default room weight for the DESTINATION room - so that the user can find
the value that they would be overriding if they enter a non-zero value
in the exit weight for a particular normal exit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>