BugFix: Some ANSI decoding errors in TBuffer class
The Factors used for converting 256 Colour codes (ANSI SGR 38;5;<X>m and
48;5;<X>m were not correct to map from 0 to 255 fro the extreme values: for
the 6x6x6 RGB cases the factor should be 51 not 42 and for the 24
Grey-scale the factor should be approximately 11 not 10.

The ANSI code SGR 39m is "Default" Foreground not Background colour, but
at around line 741 in TBuffer.cpp the code was interpreted as if it was for
a default background.  This commit fixes this and moves the original code
block to correctly work against the SGR 49m code.  A related comment
elsewhere against the same (39m) code was revised to clarify this.

In TBuffer::translateToPlainText(...) the wrong boolean flag that controls
part of the "state machine" is reset it should be
TBuffer::mHighColorModeBackground but mHighColorModeForeground was being
used.

In a couple of places code that processes ANSI SGR 38...m and 48...m
comments have been added in relation to handling values other than 5 as
the first parameter element - particularly 2 which is for direct 24-bit RGB
colour codes.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>