Update: format of modded code & +ed commented support new text format opts
Adjust code layout for agreed formatting styles (mainly trail opening '{'
on same line as text for conditional code and enforce '{''}' use on single
code line blocks).  Also move some text strings to QStringLiteral() or tr()
wrappers as appropriate.

The previous commit but one (commit-505ec100) added one more additional
text formatting option to console text, this commit adds some fragments
towards partial support for a couple of others (overline & inverse) in
commented out additional code.

Revised tool-tip message for profile preference "log in HTML" option, the
text was out of date as it referred to using the raw (non-HTML) log as
input to test the user's triggers - this functionality is now the separate
"Recording" function.  The tool-tip was not in the HTML form that auto-wrap
the text when displayed, instead it was a single unwrapped line across the
screen.  Add same HTML wrapper to logButton tooltips,

Removed TConsole::mLastBufferLogLine as it was only used by old commented
out logging code (also removed here) probably a remnant of what the
previous paragraph's old tool-tip referred to.

Fixes the loss from log files of all but the last line of long lines of
user input and other text that gets automatically wrapped to fit on the
console display.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>