Enhance: lua searchRoom() search, allow non-ASCII room names
This implements PR #213 (commit-f5b6b42e) as applied to "development"
branch to the release_30 one, it includes a squashed in second commit
to fix a small redundant bit of code in the original.

Update getRoomName(), setRoomName() and searchRoom() to handle non-ASCII
room name data.  The strings passed to and from the lua subsystem now use
Utf-8 which is a super-set of (backwards compatible with) ASCII.

Also, the original lua command when invoked as searchRoom("string") did a
case insensitive room name search and used the supplied string for a
partial search of all room names in the map.  One or two optional Boolean
arguments can be now supplement the supplied search string for the
room-name, the first, if true, makes the search case sensitive and the
second, requires the string to be a complete match rather than a sub-string
of the room name to be found.  Both of these arguments can make it easier
to find a specific room.

In the same manner as the above mentioned commit addressing map room user
data the error messages from these three lua commands have been made
capable of being translated by the Qt internationalisation system, and the
resultant texts can be passed as Utf-8 data to the lua system and to user
with non-ASCII characters.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>