RefactorAndEnhance: Adds ability to manually import an XML map file
Added code to allow manual importation of suitable XML map files that
might have been download manually.  Added method
(bool)TConsole::importMap(QString) for this. Can be activate by choosing
"Mudlet XML import file (*.xml)" on the dlgProfilePreferences dialog's map
loading control.  This now makes a reasonable job of manually importing a
usable map from say the Achaea MUD or by trans-coding from another client.
Specifically new features is that doors and hidden exits are now imported,
from door = 1 (or 2 or 3) and hidden = 1 {the same as a door = 3) in the
attributes of a <coord> sub-element of the <room> element.  The TArea
class data is now also correctly generated whereas it was being skipped
previously which left an unusable map afterwards.

Re-factoring the code in XMLimport class so that each method tends to
parse one tag and then returns for the next sibling makes it easier to
handle cases where the format has changed slightly and elements have
been added or removed.  It has also means that the code process to process
an unexpected entry (say from an item that is no longer required) can be
handled more cleanly by a generic XML::readUnknownElement() rather than a
different function for each different type of tag.

Added debug code to display ALL attributes of an element and more data
about the header of the XML file - helps to find out if things are not as
they should be!

The Host::mRoomSize and Host::mLineSize were being controlled by both
attributes in the <Host> element AND subsidiary <mRoomSize> and <mLineSize>
elements, given that the latter are more effort to parse and the problem
with the former (they were being treated as integer when they are actually
doubles and the first at least usually has a value less then 1) it is
easiest to just fix the former.

Also removed superfluous and erroneous TTextEdit::mCommandSeperator and
Host::mCommandSeperator members.  The relevant functionality should be
based on Host::mCommandSeparator

Also TRoom::auditExits() now reports via debugging output the id of normal
but invalid exit room Ids before clearing them by setting them to -1.  It
might be worth adding the data to the Room User Data for user reference.
However the current omission of a previous enhancement in that area on the
release_30 branch would counter the benefit of doing this.

Also wrapped all string in XMLimport class with QStringLiteral()/tr()
wrappers.