Bugfix/Enhance:  Suppress debug output from short interval timers
Adds a spinbox control to the last page of the profile preferences that
allows the user to set an interval for which any Timer less than this will
NOT dump it's command and script content to the debug console in dark
yellow on a blue background EVERY time it fires.  The Execute OK line will
still show up in green - so the timer can be seen to be firing but the
windows will not fill up with spammy text so much.

Also: I also spotted and corrected a couple of lines above my addition to
dlgProfilePreferences::slot_save_and_exit() where the unchecked global
pointer to the Host (mpHost) was being used instead of the local checked
pointer (pHost) which is a copy of it and used everywhere else in the
method.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>