BugFix: faulty message alignment, cTelnet::postMessage() in non-ERROR cases
A version of this commit never got into the release_30 branch with the
effect that all the postMessage() type calls with messages OTHER than the
ones beginning with "[ ERROR ]" were not being aligned correctly.

Original commit-e17d023b read:
===========================================================================
Code that was put in to fix messages with tag "[ ERROR ] - " was not copied
to the other tag types, so those messages were not lined up correctly as
per:

"[<tag>] - <message line 1 ...>
           <message line 2 ...>
           ...
           <message line n ...>"

===========================================================================

Whilst fixing this I decided to give the Host class a wrapper around
mTelnet.postMessage() so that other classes could access this without
having to include the cTelnet header or access it via a Host pointer.
As a result other existing uses have been modified where it is advantagous
to do so.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>